\name{YPmodelPhreg-internal}
\alias{YPmodelPhreg-internal}
\alias{fn_ux}
\alias{fn_uhx}
\alias{fn_uh}
\alias{fn_u}
\alias{fn_summ}
\alias{fn_hrci}
\alias{fn_est0_ypmodel0}
\alias{fn_est_ypmodel0}
\alias{fn_est_ypmodelx}
\title{Internal functions for the YPmodelPhreg package}
\description{
Internal functions for the YPmodelPhreg package.
}
\value{
No return value, called for side effects
}
